+++ src/pyrosimple/util/matching.py
@@ -350,7 +350,7 @@ class PatternFilter(FieldFilter):
             split_needle = self.SPLIT_PRE_GLOB_RE.split(needle)
         # If the cleaning of the needle did not succeed and we still
         # have regex-y values, don't prefilter
-        if any(metachar in needle for metachar in r'{}[]\\|()'):
+        if any(metachar in needle for metachar in r"{}[]\\|()"):
             return ""
         # Grab the longest needle available from the array
         needle = list(sorted(split_needle, key=len))[-1]